// src/types/metrics.ts
export interface Metrics {
  colors: number;
  mediaQueries: number;
  selectors: number;
  selectorsByAttribute: number;
  selectorsByClass: number;
  selectorsById: number;
  selectorsByPseudo: number;
  selectorsByTag: number;
  rules: number; // CSS Rules
  declarations: number; // Properties
}